app/backend/requirements.txt
flask
flask-socketio
flask-cors
python-socketio
eventlet
python-chess
SQLAlchemy
passlib[bcrypt]
PyJWT
pydantic
